Refactor: factor out push_pull.rs
Since pull_inputs and push_outputs involves a lot of code now.
