+++ src/construction.rs
@@ -1,8 +1,9 @@
 use std::{collections::HashMap, sync::OnceLock};
 
 use crate::{
-    building::{pull_inputs, push_outputs, Building, BuildingType, HasInventory},
+    building::{Building, BuildingType},
     crew::{expected_crew_deliveries, Crew},
+    push_pull::{pull_inputs, push_outputs, HasInventory},
     task::{Direction, BUILD_CONVEYOR_TIME, BUILD_POWER_GRID_TIME},
     transport::{expected_deliveries, Transport},
     Conveyor, Inventory, ItemType, Pos, WIDTH,
