+++ src/lib.rs
@@ -4,6 +4,7 @@ mod construction;
 mod conveyor;
 mod crew;
 mod info;
+mod push_pull;
 mod render;
 mod task;
 mod transport;
